Move some setup to base image + more...
Move the chain initialisation and DAG generation to the base image so
they won't be re-created each time docker compose rebulds the images.

Also switch to using the scale feature of Compose instead of recreating
the same node over and over in the yaml config. Still need to get the
nodes to connect to one another.

Add a README with some instructions.